+++ src/routes/+layout.svelte
@@ -68,7 +68,8 @@
             label: 'useInfiniteScroll',
             icon: 'lucide:arrow-down-to-line'
         },
-        { href: '/use-escape-keydown', label: 'useEscapeKeydown', icon: 'lucide:keyboard' }
+        { href: '/use-escape-keydown', label: 'useEscapeKeydown', icon: 'lucide:keyboard' },
+        { href: '/use-debounce', label: 'useDebounce', icon: 'lucide:timer' }
     ]
 
     const docItems = [